# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: cuanto_cuesta API
  version: "1.0.0"
  description: >
    Catalog of service-offering businesses (salons, barbers, spas, nail
    studios) aggregated from multiple sources. Each business is a canonical,
    merged record with per-source provenance and freshness.
servers:
  - url: /
paths:
  /health:
    get:
      summary: Liveness probe
      operationId: health
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
  /v1/businesses:
    get:
      summary: List businesses
      operationId: listBusinesses
      parameters:
        - { name: category, in: query, schema: { type: string }, description: "Category slug, e.g. barberia" }
        - { name: city, in: query, schema: { type: string }, description: "City slug, e.g. madrid" }
        - { name: q, in: query, schema: { type: string }, description: "Case-insensitive name substring" }
        - { name: min_rating, in: query, schema: { type: number, minimum: 0, maximum: 5 } }
        - { name: lat, in: query, schema: { type: number }, description: "Latitude; requires lng and radius_km" }
        - { name: lng, in: query, schema: { type: number }, description: "Longitude; requires lat and radius_km" }
        - { name: radius_km, in: query, schema: { type: number, maximum: 500 }, description: "Geo radius in km" }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: A page of businesses (summary view)
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/BusinessSummary" } }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/businesses/{id}:
    get:
      summary: Get a business (full detail)
      operationId: getBusiness
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - $ref: "#/components/parameters/IfNoneMatch"
        - $ref: "#/components/parameters/IfModifiedSince"
      responses:
        "200":
          description: Full business detail
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Last-Modified: { $ref: "#/components/headers/LastModified" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BusinessDetail" }
        "304": { description: Not modified (client copy is current) }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/businesses/{id}/services:
    get:
      summary: Get a business's service menu
      operationId: getBusinessServices
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - $ref: "#/components/parameters/IfNoneMatch"
        - $ref: "#/components/parameters/IfModifiedSince"
      responses:
        "200":
          description: The service menu
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Last-Modified: { $ref: "#/components/headers/LastModified" }
          content:
            application/json:
              schema:
                type: object
                properties:
                  business_id: { type: integer, format: int64 }
                  items: { type: array, items: { $ref: "#/components/schemas/Service" } }
                  total: { type: integer }
        "304": { description: Not modified }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/businesses/{id}/reviews:
    get:
      summary: Get a business's sample reviews
      operationId: getBusinessReviews
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - $ref: "#/components/parameters/IfNoneMatch"
        - $ref: "#/components/parameters/IfModifiedSince"
      responses:
        "200":
          description: Sample reviews (freshest per source)
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
            Last-Modified: { $ref: "#/components/headers/LastModified" }
          content:
            application/json:
              schema:
                type: object
                properties:
                  business_id: { type: integer, format: int64 }
                  items: { type: array, items: { $ref: "#/components/schemas/Review" } }
                  total: { type: integer }
        "304": { description: Not modified }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/scrape:
    post:
      summary: Start a scrape job (async, single-flight)
      operationId: startScrape
      security: [{ bearerAuth: [] }]
      description: >
        Starts a background ingest run and returns immediately. Only one job
        runs at a time; starting another while one is in progress returns 409
        with the in-flight job. Requires a bearer token; disabled (503) when
        the server has no token configured.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScrapeRequest" }
      responses:
        "202":
          description: Job accepted and running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapeJob" }
        "401": { description: Missing or invalid bearer token }
        "409":
          description: A scrape is already running
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapeJob" }
        "503": { description: Scrape endpoints disabled (no token configured) }
    get:
      summary: Status of the most recent scrape job
      operationId: scrapeStatus
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current or last job (or idle if none has run)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScrapeJob" }
        "401": { description: Missing or invalid bearer token }
        "503": { description: Scrape endpoints disabled }
  /v1/categories:
    get:
      summary: Browse index of categories with counts
      operationId: listCategories
      responses:
        "200": { $ref: "#/components/responses/FacetList" }
  /v1/cities:
    get:
      summary: Browse index of cities with counts
      operationId: listCities
      responses:
        "200": { $ref: "#/components/responses/FacetList" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema: { type: string }
      description: ETag from a prior response; a match yields 304.
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      schema: { type: string }
      description: HTTP-date from a prior Last-Modified; not-newer yields 304.
  headers:
    ETag:
      schema: { type: string }
      description: Validator derived from id and last_verified.
    LastModified:
      schema: { type: string }
      description: HTTP-date of last_verified.
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Business not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    FacetList:
      description: Distinct values with counts, most-populated first
      content:
        application/json:
          schema:
            type: object
            properties:
              items: { type: array, items: { $ref: "#/components/schemas/Facet" } }
              total: { type: integer }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Address:
      type: object
      properties:
        street: { type: string }
        locality: { type: string }
        postal_code: { type: string }
        country: { type: string }
    Location:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
    Rating:
      type: object
      properties:
        value: { type: number }
        review_count: { type: integer }
    OpeningHours:
      type: object
      properties:
        days: { type: array, items: { type: string } }
        opens: { type: string, example: "09:00" }
        closes: { type: string, example: "20:30" }
    Service:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, nullable: true }
        price_range: { type: string, description: "Set when the source prices a span (minPrice/maxPrice)" }
        currency: { type: string }
        duration_min: { type: integer, nullable: true }
        image_url: { type: string }
    Review:
      type: object
      properties:
        source: { type: string }
        author: { type: string }
        rating: { type: number, nullable: true }
        body: { type: string }
        date: { type: string, example: "2026-05-29" }
    SourceListing:
      type: object
      description: Per-source provenance for one canonical business.
      properties:
        source: { type: string }
        url: { type: string }
        rating: { $ref: "#/components/schemas/Rating" }
        price_range: { type: string }
        services: { type: integer, description: "Number of services on this listing" }
        scraped_at: { type: string, format: date-time }
    Facet:
      type: object
      properties:
        value: { type: string }
        count: { type: integer }
    ScrapeRequest:
      type: object
      description: All fields optional; the server fills sensible defaults.
      properties:
        sources:
          type: array
          items: { type: string, enum: [booksy, treatwell, web, crawl] }
          description: "Defaults to [booksy, treatwell]"
        country: { type: string, example: es }
        limit: { type: integer, description: "Max businesses per source" }
        concurrency: { type: integer }
        rps: { type: number, description: "Per-host requests/second" }
        refresh_older_than_hours: { type: number, description: "Re-crawl stale listings instead of discovering new ones" }
        seeds: { type: array, items: { type: string }, description: "Seed URLs for web/crawl sources" }
        crawl_depth: { type: integer }
        max_pages: { type: integer }
        per_host_cap: { type: integer }
    ScrapeJob:
      type: object
      properties:
        id: { type: string }
        status: { type: string, enum: [running, completed, failed] }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        options: { $ref: "#/components/schemas/ScrapeRequest" }
        result:
          type: object
          properties:
            saved: { type: integer }
            failed: { type: integer }
            by_source:
              type: object
              additionalProperties:
                type: object
                properties:
                  saved: { type: integer }
                  failed: { type: integer }
        error: { type: string }
    BusinessSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
        schema_type: { type: string }
        description: { type: string }
        city: { type: string }
        address: { $ref: "#/components/schemas/Address" }
        location: { $ref: "#/components/schemas/Location" }
        rating: { $ref: "#/components/schemas/Rating" }
        price_range: { type: string }
        price_from: { type: number }
        price_to: { type: number }
        price_currency: { type: string }
        distance_km: { type: number, description: "Present only on geo-filtered lists" }
        image_url: { type: string }
        logo_url: { type: string }
        sponsored: { type: boolean, description: "Associated/partner business; partners sort first in the default list (demo: deterministic per id)" }
        verified: { type: boolean, description: "Identity/legitimacy confirmed (demo: deterministic per id)" }
        sources: { type: array, items: { type: string } }
        last_verified: { type: string, format: date-time }
        stale: { type: boolean, description: "true when last_verified is older than 30 days" }
    BusinessDetail:
      allOf:
        - $ref: "#/components/schemas/BusinessSummary"
        - type: object
          properties:
            phone: { type: string }
            email: { type: string }
            payment_accepted: { type: string }
            images: { type: array, items: { type: string } }
            social_links: { type: array, items: { type: string } }
            opening_hours: { type: array, items: { $ref: "#/components/schemas/OpeningHours" } }
            services: { type: array, items: { $ref: "#/components/schemas/Service" } }
            reviews: { type: array, items: { $ref: "#/components/schemas/Review" } }
            listings: { type: array, items: { $ref: "#/components/schemas/SourceListing" } }